requirements.txt
django>=1.7
django-tastypie>=0.10